chore(backend): retire screenalytics runtime surfaces
