+++ .env.example
@@ -197,16 +197,16 @@ SUPABASE_HTTP_POOL_TIMEOUT_SEC=5
 # ----------------------------
 # Object Storage
 # ----------------------------
-# Provider-neutral S3-compatible object storage contract.
-OBJECT_STORAGE_PROVIDER=
-OBJECT_STORAGE_BUCKET=
-OBJECT_STORAGE_REGION=
-OBJECT_STORAGE_ENDPOINT_URL=
+# Canonical backend object-storage contract. Production uses Cloudflare R2.
+OBJECT_STORAGE_PROVIDER=r2
+OBJECT_STORAGE_BUCKET=trr-media-prod
+OBJECT_STORAGE_REGION=auto
+OBJECT_STORAGE_ENDPOINT_URL=https://<ACCOUNT_ID>.r2.cloudflarestorage.com
 OBJECT_STORAGE_ACCESS_KEY_ID=
 OBJECT_STORAGE_SECRET_ACCESS_KEY=
 OBJECT_STORAGE_SESSION_TOKEN=
 OBJECT_STORAGE_PROFILE=
-OBJECT_STORAGE_PUBLIC_BASE_URL=
+OBJECT_STORAGE_PUBLIC_BASE_URL=https://media.thereality.report
 OBJECT_STORAGE_PREFIX=
 
 # ----------------------------
@@ -236,15 +236,6 @@ TRR_INTERNAL_ADMIN_SHARED_SECRET=
 
 # TRR API URL (this service)
 TRR_API_URL=http://localhost:8000
-
-# Screenalytics API base URL (optional legacy outbound HTTP callers only).
-# Production screentime no longer depends on this setting.
-# When running screenalytics alongside TRR-Backend locally, use a non-:8000 port (workspace default is :8001).
-SCREENALYTICS_API_URL=http://localhost:8001
-
-# Optional legacy Screenalytics service-to-service auth.
-# Production screentime no longer requires this token.
-SCREENALYTICS_SERVICE_TOKEN=
 
 # Better Stack / Logtail HTTP ingest for backend and Modal job logs.
 # Recommended for the Render cutover steady state; default operator choice is Better Stack free first.
