+++ api/screenalytics_auth.py
@@ -1,10 +1,8 @@
-"""Screenalytics service-to-service auth."""
+"""Auth shim for retained internal screentime worker endpoints."""
 
 from __future__ import annotations
 
-import hmac
 import logging
-import os
 
 from fastapi import HTTPException, Request
 
@@ -12,22 +10,6 @@ from api.auth import get_bearer_token
 from trr_backend.security.internal_admin import InvalidTokenError, verify_internal_admin_token
 
 logger = logging.getLogger(__name__)
-
-
-def get_screenalytics_service_token() -> str | None:
-    token = os.getenv("SCREENALYTICS_SERVICE_TOKEN", "").strip()
-    return token or None
-
-
-def _internal_admin_secret_configured() -> bool:
-    return bool((os.getenv("TRR_INTERNAL_ADMIN_SHARED_SECRET") or "").strip())
-
-
-def _env_flag(name: str, default: bool = False) -> bool:
-    raw = (os.getenv(name) or "").strip().lower()
-    if not raw:
-        return default
-    return raw not in {"0", "false", "no", "off"}
 
 
 async def require_screenalytics_service_token(request: Request) -> None:
@@ -39,22 +21,9 @@ async def require_screenalytics_service_token(request: Request) -> None:
             headers={"WWW-Authenticate": "Bearer"},
         )
 
-    expected = get_screenalytics_service_token()
-    allow_service_token_fallback = _env_flag("TRR_SCREENALYTICS_ALLOW_SERVICE_TOKEN_FALLBACK", False)
-
-    if expected and allow_service_token_fallback and hmac.compare_digest(token, expected):
-        logger.info("screenalytics auth fallback accepted via service token")
-        return None
-
-    if expected and allow_service_token_fallback and not _internal_admin_secret_configured():
-        raise HTTPException(
-            status_code=401,
-            detail="Invalid service token",
-            headers={"WWW-Authenticate": "Bearer"},
-        )
-
     try:
         verify_internal_admin_token(token)
+        logger.debug("retained screentime worker request accepted via internal admin token")
         return None
     except RuntimeError as exc:
         raise HTTPException(
@@ -63,17 +32,10 @@ async def require_screenalytics_service_token(request: Request) -> None:
             headers={"x-error-code": "AUTH_SERVICE_UNAVAILABLE"},
         ) from exc
     except InvalidTokenError:
-        pass
-
-    if expected and allow_service_token_fallback:
-        raise HTTPException(
-            status_code=401,
-            detail="Invalid service token",
-            headers={"WWW-Authenticate": "Bearer"},
-        )
+        logger.info("retained screentime worker request rejected: invalid internal admin token")
 
     raise HTTPException(
         status_code=401,
-        detail="Valid internal admin token required; service-token fallback is transitional and disabled by default",
+        detail="Valid internal admin token required for retained screentime worker endpoints",
         headers={"WWW-Authenticate": "Bearer"},
     )
