+++ docs/ai/HANDOFF.md
@@ -5,7 +5,7 @@ Generated by `scripts/sync-handoffs.py`. Do not edit by hand.
 Purpose: active-work index for multi-turn AI agent sessions in `TRR-Backend`. Keep this file short.
 
 ## Current Active Work
-- `TASK28 Concerns remediation and Screenalytics contract lock` | last updated `2026-04-09` | current phase `backend contract cutover landed` | next action `continue backend decomposition and remove remaining Screenalytics compatibility surfaces` | details `../cross-collab/TASK28/STATUS.md`
+- None.
 
 ## Blocked / Waiting
 - None.
@@ -16,6 +16,7 @@ Purpose: active-work index for multi-turn AI agent sessions in `TRR-Backend`. Ke
 - `Modal Instagram Auth Repair` | last updated `2026-04-06` | current phase `modal instagram auth repair` | next action `Refresh the admin social profile page and run Backfill Posts.` | details `local-status/modal-instagram-auth-repair.md`
 
 ## Older Plans
+- `TASK28 Concerns remediation and Screenalytics contract lock` | last updated `2026-04-09` | current phase `backend contract cutover landed` | next action `continue backend decomposition and remove remaining Screenalytics compatibility surfaces` | details `../cross-collab/TASK28/STATUS.md`
 - `TASK27 Instagram Backfill Worker Reliability` | last updated `2026-04-07` | current phase `running targeted verification for the backend reliability changes` | next action `finish validation, update the runbook/status evidence, and close out the backend implementation summary` | details `../cross-collab/TASK27/STATUS.md`
 - `TASK26 Instagram Shared-Profile Rollout Guardrails` | last updated `2026-04-04` | current phase `frontier/auth fix proven live; instagram session repair pending` | next action `repair the Instagram authenticated session, then rerun bounded canary and affected replay against the deployed backend` | details `../cross-collab/TASK26/STATUS.md`
 - `TASK24 Final Supabase connection audit and donor transition inventory` | last updated `2026-04-03` | current phase `phase 5 screentime runtime retirement implemented` | next action `treat the inventory as historical donor context; active screentime runtime dependency is retired` | details `../cross-collab/TASK24/STATUS.md`
@@ -25,7 +26,6 @@ Purpose: active-work index for multi-turn AI agent sessions in `TRR-Backend`. Ke
 - `Cast Screentime Phase 4 Review Publication Cutover` | last updated `2026-04-03` | current phase `phase 4 review and publication cutover implemented` | next action `Plan Phase 5 runtime retirement and remove the remaining rollback-only Screenalytics boundary.` | details `local-status/cast-screentime-phase4-review-publication-cutover.md`
 - `Cast Screentime Phase 3 Backend Execution Port` | last updated `2026-04-03` | current phase `phase 3 backend execution port implemented` | next action `Plan Phase 4 review/publication cutover against the retained backend runtime.` | details `local-status/cast-screentime-phase3-backend-execution-port.md`
 - `Cast Screentime Phase 2 Identity Reset` | last updated `2026-04-03` | current phase `phase 2 identity reset implemented` | next action `port retained screentime execution onto the reviewed face-reference and DeepFace embedding contract` | details `local-status/cast-screentime-phase2-identity-reset.md`
-- `TASK25 TRR-APP Supabase Simplification and Index Hardening` | last updated `2026-04-02` | current phase `backend migration complete` | next action `keep the generated schema docs with the migration and carry the recorded verification caveats forward` | details `../cross-collab/TASK25/STATUS.md`
 
 ## Archives / Canonical Links
 - History archive: `archive/HANDOFF-legacy-2026-03-16.md`
