+++ docs/api/run.md
@@ -23,15 +23,6 @@ The API requires the following environment variables to be set:
 | `SUPABASE_ANON_KEY` | Optional anon key for non-admin API flows | `eyJ...` |
 | `SUPABASE_SERVICE_ROLE_KEY` | Service role key for backend-owned Supabase admin operations | `eyJ...` |
 
-### Screenalytics compatibility (Optional legacy)
-
-| Variable | Description | Example |
-|----------|-------------|---------|
-| `SCREENALYTICS_SERVICE_TOKEN` | Optional legacy bearer token for old `/api/v1/screenalytics/*` callers | `change-me-long-random-token` |
-| `SCREENALYTICS_API_URL` | Optional base URL for legacy outbound Screenalytics HTTP callers | `https://screenalytics.example.com` |
-
-Production screentime flows no longer require `SCREENALYTICS_SERVICE_TOKEN` or `SCREENALYTICS_API_URL`. Legacy `/api/v1/screenalytics/*` and `/api/v1/screenalytics/v2/*` endpoints now default to signed internal admin JWT auth; service-token fallback is transitional and must be explicitly re-enabled with `TRR_SCREENALYTICS_ALLOW_SERVICE_TOKEN_FALLBACK=1`. Keep `SCREENALYTICS_API_URL` only if you still run explicit legacy outbound HTTP calls to the standalone Screenalytics service.
-
 ### Internal Admin Proxy (TRR-APP -> TRR-Backend)
 
 | Variable | Description | Example |
@@ -93,7 +84,7 @@ When `BETTER_STACK_SOURCE_TOKEN` is set, the backend ships structured Python log
 | Variable | Description | Example |
 |----------|-------------|---------|
 | `OBJECT_STORAGE_PROVIDER` | Storage backend name for operator clarity | `r2` |
-| `OBJECT_STORAGE_BUCKET` | Canonical object-storage bucket name | `trr-backend` |
+| `OBJECT_STORAGE_BUCKET` | Canonical object-storage bucket name | `trr-media-prod` |
 | `OBJECT_STORAGE_REGION` | Region or provider region token | `auto` |
 | `OBJECT_STORAGE_ENDPOINT_URL` | S3-compatible API endpoint | `https://<accountid>.r2.cloudflarestorage.com` |
 | `OBJECT_STORAGE_ACCESS_KEY_ID` | S3-compatible access key | `...` |
@@ -171,7 +162,6 @@ FastAPI ships with interactive API docs and an OpenAPI schema:
 | --- | --- | --- |
 | TRR App | `Authorization: Bearer <Supabase access token>` | Use for user-scoped endpoints under `/api/v1/*`. |
 | TRR App internal proxy (facebank toggle) | `Authorization: Bearer <internal admin JWT>` | Allowed only for `PATCH /api/v1/admin/person/{person_id}/gallery/{link_id}/facebank-seed`; TRR-APP signs the JWT with `TRR_INTERNAL_ADMIN_SHARED_SECRET`. |
-| Legacy Screenalytics compatibility | Signed internal admin JWT by default; optional `Authorization: Bearer <SCREENALYTICS_SERVICE_TOKEN>` only when `TRR_SCREENALYTICS_ALLOW_SERVICE_TOKEN_FALLBACK=1` | `/api/v1/screenalytics/*` and `/api/v1/screenalytics/v2/*` remain compatibility routes; production screentime does not require them. |
 
 Admin allowlist
 - Facebank seed toggle endpoint requires either an allowlisted user JWT or a valid internal admin JWT signed with `TRR_INTERNAL_ADMIN_SHARED_SECRET`.
@@ -189,9 +179,9 @@ Admin allowlist
 curl -H "Authorization: Bearer $SUPABASE_ACCESS_TOKEN" \
   http://localhost:8000/api/v1/shows
 
-# Legacy Screenalytics compatibility (service token if still configured)
-curl -H "Authorization: Bearer $SCREENALYTICS_SERVICE_TOKEN" \
-  http://localhost:8000/api/v1/screenalytics/episodes/<episode_id>/cast
+# Internal admin JWT (TRR-APP proxy pattern)
+curl -H "Authorization: Bearer $TRR_INTERNAL_ADMIN_JWT" \
+  http://localhost:8000/api/v1/admin/person/<person_id>/gallery/<link_id>/facebank-seed
 ```
 
 ## Running Tests
